Update files, Alura, Introdução a C, Aula 3.3

+++ alura/c/adivinhacao.c
@@ -12,7 +12,7 @@ int main() {
   int chute;
   int tentativas = 1;
 
-  int pontos = 1000;
+  double pontos = 1000;
 
   while(1) {
 
@@ -47,13 +47,13 @@ int main() {
 
     tentativas++;
 
-    int pontosperdidos = (chute - numerosecreto) / 2;
+    double pontosperdidos = (chute - numerosecreto) / 2.0;
     pontos = pontos - pontosperdidos;
 
   }
 
   printf("Fim de jogo!\n");
   printf("Você acertou em %d tentativas!\n", tentativas);
-  printf("Total de pontos: %d\n", pontos);
+  printf("Total de pontos: %f\n", pontos);
 
 }
